Unify page bootstrapping
